+++ Stitch/Graph/Node/Port/View/Field/InputView/LayerGroupOrientationDropDownChoiceView.swift
@@ -152,231 +152,3 @@ extension PortValue {
         }
     }
 }
-
-
-// this UI is quite a bit different, because you are receiving a full PortValue.anchoring enum,
-// but want to show something more specific
-extension Anchoring {
-    func sfSymbolForLayerGroupOrientation(_ orientation: StitchOrientation) -> String {
-        switch orientation {
-        case .grid, .none:
-            fatalErrorIfDebug()
-            return "" // should not happen
-        case .horizontal:
-            return "align.horizontal.left"
-        case .vertical:
-            return "align.vertical.top"
-        }
-    }
-}
-
-enum HorizontalAlignmentChoices: Equatable, Hashable, CaseIterable {
-    case left, center, right
-    
-    var sfSymbol: String {
-        switch self {
-        case .left:
-            return "align.vertical.left"
-        case .center:
-            return "align.vertical.center"
-        case .right:
-            return "align.vertical.right"
-        }
-    }
-    
-    var asAnchoring: Anchoring {
-        switch self {
-        case .left:
-            return .centerLeft
-        case .center:
-            return .centerCenter
-        case .right:
-            return .centerRight
-        }
-    }
-    
-    static func fromAnchoring(_ anchoring: Anchoring) -> Self {
-        switch anchoring {
-        case .topLeft:
-            return .left
-        case .topCenter:
-            return .center
-        case .topRight:
-            return .right
-        case .centerLeft:
-            return .left
-        case .centerCenter:
-            return .center
-        case .centerRight:
-            return .right
-        case .bottomLeft:
-            return .left
-        case .bottomCenter:
-            return .center
-        case .bottomRight:
-            return .right
-        
-        /*
-         TODO: turn various numbers into proper alignments? e.g.
-         x < 0.3 is .leading alignment,
-         x > 0.7 is .trailing;
-         everything else is .center
-         */
-        default:
-            return .center // some default
-        }
-    }
-}
-
-enum VerticalAlignmentChoices: Equatable, Hashable, CaseIterable {
-    case top, center, bottom
-    
-    var sfSymbol: String {
-        switch self {
-        case .top:
-            return "align.vertical.top"
-        case .center:
-            return "align.vertical.center"
-        case .bottom:
-            return "align.vertical.bottom"
-        }
-    }
-    
-    var asAnchoring: Anchoring {
-        switch self {
-        case .top:
-            return .topCenter
-        case .center:
-            return .centerCenter
-        case .bottom:
-            return .bottomCenter
-        }
-    }
-    
-    static func fromAnchoring(_ anchoring: Anchoring) -> Self {
-        switch anchoring {
-        case .topLeft:
-            return .top
-        case .topCenter:
-            return .top
-        case .topRight:
-            return .top
-        case .centerLeft:
-            return .center
-        case .centerCenter:
-            return .center
-        case .centerRight:
-            return .center
-        case .bottomLeft:
-            return .bottom
-        case .bottomCenter:
-            return .bottom
-        case .bottomRight:
-            return .bottom
-        
-        /*
-         TODO: turn various numbers into proper alignments? e.g.
-         y < 0.3 is .top alignment,
-         y > 0.7 is .bottom;
-         everything else is .center
-         */
-        default:
-            return .center
-        }
-    }
-}
-
-
-struct LayerGroupVerticalAlignmentPickerFieldValueView: View {
-    @State var currentChoice: VerticalAlignmentChoices = .center
-    
-    let id: InputCoordinate
-    let value: Anchoring
-    
-    // VStack = show ...
-    // HStack = show ...
-    // None or Grid = this field will actually be blocked
-    
-//    let layerGroupOrientation: StitchOrientation
-    
-    var isVStack: Bool {
-        true
-    }
-    
-    let layerInputObserver: LayerInputObserver?
-    let isFieldInsideLayerInspector: Bool
-    
-    @MainActor
-    var hasHeterogenousValues: Bool {
-        if let layerInputObserver = layerInputObserver {
-            @Bindable var layerInputObserver = layerInputObserver
-            return layerInputObserver.fieldHasHeterogenousValues(
-                0,
-                isFieldInsideLayerInspector: isFieldInsideLayerInspector)
-        } else {
-            return false
-        }
-    }
-    
-    var body: some View {
-        Picker("", selection: $currentChoice) {
-            ForEach(VerticalAlignmentChoices.allCases, id: \.self) { choice in
-                Image(systemName: choice.sfSymbol).tag(choice)
-            }
-        }
-        .pickerStyle(.segmented)
-        .scaledToFit()
-        .frame(width: 148, height: NODE_ROW_HEIGHT * 2, alignment: .trailing)
-        .onChange(of: self.currentChoice) { oldValue, newValue in
-                dispatch(PickerOptionSelected(
-                    input: self.id,
-                    choice: .anchoring(newValue.asAnchoring),
-                    isFieldInsideLayerInspector: isFieldInsideLayerInspector))
-        }
-        .onAppear {
-            self.currentChoice = .fromAnchoring(value)
-        }
-    }
-}
-
-struct SpecialPickerFieldValueView: View {
-    @State var currentChoice: PortValue
-    
-    let id: InputCoordinate
-    let value: PortValue
-    let choices: PortValues
-    let layerInputObserver: LayerInputObserver?
-    let isFieldInsideLayerInspector: Bool
-    
-    @MainActor
-    var hasHeterogenousValues: Bool {
-        if let layerInputObserver = layerInputObserver {
-            @Bindable var layerInputObserver = layerInputObserver
-            return layerInputObserver.fieldHasHeterogenousValues(
-                0,
-                isFieldInsideLayerInspector: isFieldInsideLayerInspector)
-        } else {
-            return false
-        }
-    }
-    
-    var body: some View {
-        Picker("", selection: $currentChoice) {
-            ForEach(choices, id: \.self) { choice in
-                Image(systemName: choice.sfSymbol).tag(choice)
-            }
-        }
-        .pickerStyle(.segmented)
-        .scaledToFit()
-        .frame(width: 148, height: NODE_ROW_HEIGHT * 2, alignment: .trailing)
-        .onChange(of: self.currentChoice) { oldValue, newValue in
-                dispatch(PickerOptionSelected(
-                    input: self.id,
-                    choice: newValue,
-                    isFieldInsideLayerInspector: isFieldInsideLayerInspector))
-        }
-        .onAppear {
-            self.currentChoice = value
-        }
-    }
-}
